// src/app/types.d.ts
export interface IServiceItem {
  icon: ReactNode;
  title: string;
  description: string;
  linkHref: string;
}

export interface IFaqItem {
  question: string;
  answer: string;
}

export interface IBenefitItem {
  icon: ReactNode;
  title: string;
  description: string;
}

export interface IPlanOfActionItem {
  title: string;
  description: string;
}